Eliminate ia64 compiler warnings.  Fix ia64 gas testsuite again. 	*  elfxx-ia64.c (get_dyn_sym_info): Cast %p argument to void *. 	* config/tc-ia64.h (ia64_init): Add prototype. 	* gas/ia64/dv-imply.d, gas/ia64/dv-mutex.d, gas/ia64/dv-safe.d, 	gas/ia64/dv-srlz.d, gas/ia64/opc-m.d: Update. 	* ia64-dis.c (print_insn_ia64): Cast away const on ia64_free_opcode 	argument. 	* ia64_gen.c (insert_deplist): Cast sizeof result to int. 	(print_dependency_table): Print NULL if semantics field not set. 	(insert_opcode_dependencies): Mark cmp parameter as unused. 	(print_main_table): Use fprintf_vma to print long long fields. 	(main): Mark argv paramter as unused.  Convert to old style definition. 	* ia64-opc.c (ia64_find_dependency): Cast sizeof result to int. 	* ia64-asmtab.c: Regnerate.

+++ opcodes/ia64-opc.c
@@ -740,7 +740,8 @@ ia64_find_dependency (index)
 {
   index = DEP(index);
 
-  if (index < 0 || index >= sizeof(dependencies) / sizeof(dependencies[0]))
+  if (index < 0
+      || index >= (int)(sizeof(dependencies) / sizeof(dependencies[0])))
     return NULL;
 
   return &dependencies[index];
